Change the way the CI is triggered
Before this commit the CI was triggered on any push but not on
pull requests.

With this commit, the CI gets triggered for pull requests and for
direct pushes but only to the main branch.

The purpose of this change is to make possible and favour contributions
through pull requests.