Bump version from 19 to 20.0
This commit introduces a new versionning scheme: for this package, its
major version will be updated everytime a more recent version
of the OCaml compiler is added. The minor version number is used as
a patch level, i.e. if bug fixes are needed although no new version
of the OCaml compiler has been released.

is supported.