stdcompat__native.ml_{byte,native}: store the generated files
These files do not contain anything that depends on the configuration stage. They thus do not need to be processed by configure and it is better to just store the result of the processing, which is the source file itself.

This commit also adds a missing end-of-line ot the end of both files.