deconf: list of files that do not need to be configured
This commit adds the deconf file that contains the list of
files which do not need to be configured, i.e. processed by
either config.status or stdcompatpp.

This list has been obtained by first configuring the tree and then
running the following command:

for f in *.in; do if diff $f `basename $f .in` > /dev/null 2>&1; then echo $f >> deconf; fi; done

In other words: files that are left unchanged by configure do not need to
be preprocessed.

This finds a list of 58 files that do actually not need to be
preprocessed.