Remove two no longer used shell scripts
Both of them were called generate_interfaces.sh, one was at the toplevel
of the source tree, the other in the interface_generator directory.

These scripts are not only no longer used, the fact that they have the
same name is also confusing so let's get rid of them completely for the time being.